refactor: ShortestPath 중복 되는 코드 추상 클래스로 이동
