+++ src/main/java/nextstep/path/domain/ShortestDistancePath.java
@@ -6,8 +6,6 @@ import org.jgrapht.graph.DefaultWeightedEdge;
 import org.jgrapht.graph.WeightedMultigraph;
 
 import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 public class ShortestDistancePath extends ShortestPath {
 
@@ -44,14 +42,5 @@ public class ShortestDistancePath extends ShortestPath {
                 .stream()
                 .mapToInt(edge -> getSectionByEdge(edge).getDuration())
                 .sum();
-    }
-
-    @Override
-    public Set<Section> getUsedSections(Station start, Station end) {
-        return findShortestPath(start, end)
-                .getEdgeList()
-                .stream()
-                .map(this::getSectionByEdge)
-                .collect(Collectors.toSet());
     }
 }
